Chek in Bug Details display/notecard

+++ buglog.client/src/services/BugService.js
@@ -1,7 +1,7 @@
 import { AppState } from '../AppState.js'
 import { logger } from '../utils/Logger'
 import { api } from './AxiosService'
-
+import Pop from '../utils/Notifier'
 class BugService {
   async getAll() {
     const res = await api.get('api/bugs')
@@ -11,28 +11,42 @@ class BugService {
 
   async getById(id) {
     const res = await api.get('api/bugs/' + id)
-    // logger.log('A single bug : ', res.data)
+    logger.log('A single bug : ', res.data)
     AppState.activebug = res.data
+  }
+
+  async getNotesByBugId(id) {
+    const res = await api.get('api/bugs/' + id + '/notes')
+    logger.log('fetched Notes by Bug Id', res.data)
+    AppState.notes = res.data
   }
 
   async create(body) {
     // debugger
     const res = await api.post('api/bugs/', body)
-    // logger.log('Added A project : ', res.data)
+    logger.log('Added A Bug : ', res.data)
     AppState.bugs.push(res.data)
     return res.data.id
   }
 
   async update(id, body) {
-    const res = await api.put('api/bugs/' + id, body)
-    logger.log('Udated Bug', res.data)
-    AppState.bugs = res.data
+    const bug = await this.getById(id)
+    if (user.id === bug.creatorId.toString()) {
+      const res = await api.put('api/bugs/' + id, body)
+      logger.log('Udated Bug', res.data)
+      AppState.bugs = res.data
+    }
   }
 
   async destroy(id) {
-    await api.delete('api/bugs/' + id)
-    AppState.bugs = AppState.bugs.filter(n => n.id !== id)
-    logger.log('Deleted Successfully')
+    if (await Pop.confirm()) {
+      const bug = await this.getById(id)
+      if (user.id === bug.creatorId.toString()) {
+        await api.delete('api/bugs/' + id)
+        AppState.bugs = AppState.bugs.filter(n => n.id !== id)
+        logger.log('Deleted Successfully')
+      }
+    }
   }
 }
 export const bugService = new BugService()
